pyproject.toml
[project]
name = "mau_server"
version = "2.0"
description = "A web server with extra things for Mau engine."
readme = "README.md"
requires-python = ">=3.12"
license = "GPL-3.0-or-later"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "mau>=2.0.1",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis>=6.0.0",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "nox>=2025.2.9",
]

[tool.uv.sources]
mau = { git = "https://github.com/miroqru/mauno" }

# Ruff linter ----------------------------------------------------------

[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "D", "N", "PL", "UP", "ANN", "W", "ASYNC"]
ignore = ["D407", "D107", "D213", "D203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
